+++ groucho-lsh-rest/src/main/java/it/cnr/iasi/saks/groucho/lsh/service/impl/IsunknownApiImpl.java
@@ -1,4 +1,4 @@
-package it.cnr.iasi.saks.groucho.lsh.rest.impl;
+package it.cnr.iasi.saks.groucho.lsh.service.impl;
 
 import it.cnr.iasi.saks.groucho.lsh.StateObserver;
 import it.cnr.iasi.saks.groucho.lsh.exceptions.LSHException;
@@ -11,26 +11,27 @@ import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
 @Service
 @Slf4j
-public class MarkApiImpl {
+public class IsunknownApiImpl {
     private StateObserverFactory stateObserverFactory = new StateObserverFactory();
     private StateObserver stateObserver = stateObserverFactory.getStateObserver();
     // TODO next StateObserver -> StateObserverLSH
     private StateObserverLSH stateObserverLSH = (StateObserverLSH) stateObserverFactory.getStateObserverLSH();
 
-    public ResponseEntity<Boolean> markState(String body) {
+    public ResponseEntity<Boolean> isStateUnknown(String body) {
         // TODO impl instead
-        // TODO if 400 error return httpstatus
+        // TODO if 400 error return httpstatus4
+        boolean state;
         try {
-            stateObserver.markState(body);
+            state = stateObserver.isStateUnknown(body);
         } catch (LSHException ex) {
             log.error("ERROR: ", ex);
             // TODO 400 || 404 checkLSHUtil return boolean
-            return new ResponseEntity(Boolean.FALSE, HttpStatus.NOT_FOUND);
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity(Boolean.TRUE, HttpStatus.OK);
+        return new ResponseEntity(new Boolean(state), HttpStatus.OK);
     }
 
-    public ResponseEntity<Boolean> markStateLSH(String stateStringLSH) {
+    public ResponseEntity<Boolean> isStateUnknownLSH(String stateStringLSH) {
         NotImplementedException ex = new NotImplementedException();
         log.error("ERROR: ", ex);
         throw ex;
